running a bash script 37

+++ .circleci/deploy.sh
@@ -6,15 +6,13 @@ echo "Pulling a image"
 docker pull 764941139426.dkr.ecr.ap-south-1.amazonaws.com/temp:latest
 echo "Pull complete"
 
-if [ "$( docker container inspect -f '{{.State.Status}}' 764941139426.dkr.ecr.ap-south-1.amazonaws.com/temp:latest )" == "running" ];
+CONTAINER_NAME='764941139426.dkr.ecr.ap-south-1.amazonaws.com/temp:latest'
 
-then
-echo "container running"
-docker stop 764941139426.dkr.ecr.ap-south-1.amazonaws.com/temp:latest
-docker rm 764941139426.dkr.ecr.ap-south-1.amazonaws.com/temp:latest
-
+if [ "${CID}" ]; then
+  echo "Container exists"
+  docker stop 764941139426.dkr.ecr.ap-south-1.amazonaws.com/temp:latest
+  docker rm 764941139426.dkr.ecr.ap-south-1.amazonaws.com/temp:latest
 fi
-
 
 # docker stop $AWS_ECR_ACCOUNT_URL/temp
 # docker rm $AWS_ECR_ACCOUNT_URL/temp
